+++ src/models/json-api.model.ts
@@ -1,42 +1,74 @@
 import find from 'lodash-es/find';
 import includes from 'lodash-es/includes';
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 import { JsonApiDatastore, ModelType } from '../services/json-api-datastore.service';
 import { ModelConfig } from '../interfaces/model-config.interface';
 import * as _ from 'lodash';
 import { AttributeMetadata } from '../constants/symbols';
+import { HttpHeaders } from '@angular/common/http';
+
+/**
+ * HACK/FIXME:
+ * Type 'symbol' cannot be used as an index type.
+ * TypeScript 2.9.x
+ * See https://github.com/Microsoft/TypeScript/issues/24587.
+ */
+// tslint:disable-next-line:variable-name
+const AttributeMetadataIndex: string = AttributeMetadata as any;
 
 export class JsonApiModel {
   id: string;
+  public modelInitialization: boolean = false;
+
   [key: string]: any;
 
-  // tslint:disable-next-line:variable-name
-  constructor(private _datastore: JsonApiDatastore, data?: any) {
+  lastSyncModels: Array<any>;
+
+  constructor(private internalDatastore: JsonApiDatastore, data?: any) {
     if (data) {
+      this.modelInitialization = true;
       this.id = data.id;
       Object.assign(this, data.attributes);
+      this.modelInitialization = false;
     }
   }
 
-  syncRelationships(data: any, included: any, level: number): void {
+  public isModelInitialization(): boolean {
+    return this.modelInitialization;
+  }
+
+  public syncRelationships(data: any, included: any, remainingModels?: Array<any>): void {
+    if (this.lastSyncModels === included) {
+      return;
+    }
+
     if (data) {
-      this.parseHasMany(data, included, level);
-      this.parseBelongsTo(data, included, level);
+      let modelsForProcessing = remainingModels;
+
+      if (!modelsForProcessing) {
+        modelsForProcessing = [].concat(included);
+      }
+
+      this.parseHasMany(data, included, modelsForProcessing);
+      this.parseBelongsTo(data, included, modelsForProcessing);
     }
+
+    this.lastSyncModels = included;
   }
 
-  save(params?: any, headers?: Headers): Observable<this> {
-    const attributesMetadata: any = this[AttributeMetadata];
-    return this._datastore.saveRecord(attributesMetadata, this, params, headers);
+  public save(params?: any, headers?: HttpHeaders, customUrl?: string): Observable<this> {
+    this.checkChanges();
+    const attributesMetadata: any = this[AttributeMetadataIndex];
+    return this.internalDatastore.saveRecord(attributesMetadata, this, params, headers, customUrl);
   }
 
   get hasDirtyAttributes() {
-    const attributesMetadata: any = this[AttributeMetadata];
+    this.checkChanges();
+    const attributesMetadata: any = this[AttributeMetadataIndex];
     let hasDirtyAttributes = false;
     for (const propertyName in attributesMetadata) {
       if (attributesMetadata.hasOwnProperty(propertyName)) {
         const metadata: any = attributesMetadata[propertyName];
-
         if (metadata.hasDirtyAttributes) {
           hasDirtyAttributes = true;
           break;
@@ -46,32 +78,42 @@ export class JsonApiModel {
     return hasDirtyAttributes;
   }
 
-  rollbackAttributes(): void {
+  private checkChanges() {
     const attributesMetadata: any = this[AttributeMetadata];
-    let metadata: any;
     for (const propertyName in attributesMetadata) {
       if (attributesMetadata.hasOwnProperty(propertyName)) {
-        if (attributesMetadata[propertyName].hasDirtyAttributes) {
-          this[propertyName] = attributesMetadata[propertyName].oldValue;
-          metadata = {
-            hasDirtyAttributes: false,
-            newValue: attributesMetadata[propertyName].oldValue,
-            oldValue: undefined
-          };
-          attributesMetadata[propertyName] = metadata;
+        const metadata: any = attributesMetadata[propertyName];
+        if (metadata.nested) {
+          this[AttributeMetadata][propertyName].hasDirtyAttributes = !_.isEqual(
+            attributesMetadata[propertyName].oldValue,
+            attributesMetadata[propertyName].newValue
+          );
+          this[AttributeMetadata][propertyName].serialisationValue = attributesMetadata[propertyName].converter(
+            Reflect.getMetadata('design:type', this, propertyName),
+            _.cloneDeep(attributesMetadata[propertyName].newValue),
+            true
+          );
         }
       }
     }
+  }
 
-    this[AttributeMetadata] = attributesMetadata;
+  public rollbackAttributes(): void {
+    const attributesMetadata: any = this[AttributeMetadataIndex];
+    for (const propertyName in attributesMetadata) {
+      if (attributesMetadata.hasOwnProperty(propertyName)) {
+        if (attributesMetadata[propertyName].hasDirtyAttributes) {
+          this[propertyName] = _.cloneDeep(attributesMetadata[propertyName].oldValue);
+        }
+      }
+    }
   }
 
   get modelConfig(): ModelConfig {
     return Reflect.getMetadata('JsonApiModelConfig', this.constructor);
   }
 
-
-  private parseHasMany(data: any, included: any, level: number): void {
+  private parseHasMany(data: any, included: any, remainingModels: Array<any>): void {
     const hasMany: any = Reflect.getMetadata('HasMany', this);
 
     if (hasMany) {
@@ -88,16 +130,22 @@ export class JsonApiModel {
             if (!includes(modelTypesFetched, typeName)) {
               modelTypesFetched.push(typeName);
               // tslint:disable-next-line:max-line-length
-              const modelType: ModelType<this> = Reflect.getMetadata('JsonApiDatastoreConfig', this._datastore.constructor).models[typeName];
+              const modelType: ModelType<this> = Reflect.getMetadata('JsonApiDatastoreConfig', this.internalDatastore.constructor).models[typeName];
 
               if (modelType) {
-                // tslint:disable-next-line:max-line-length
-                const relationshipModels: JsonApiModel[] = this.getHasManyRelationship(modelType, relationship.data, included, typeName, level);
+                const relationshipModels: JsonApiModel[] = this.getHasManyRelationship(
+                  modelType,
+                  relationship.data,
+                  included,
+                  typeName,
+                  remainingModels
+                );
+
                 if (relationshipModels.length > 0) {
                   allModels = allModels.concat(relationshipModels);
                 }
               } else {
-                throw { message: 'parseHasMany - Model type for relationship ' + typeName + ' not found.' };
+                throw { message: `parseHasMany - Model type for relationship ${typeName} not found.` };
               }
             }
 
@@ -110,7 +158,7 @@ export class JsonApiModel {
     }
   }
 
-  private parseBelongsTo(data: any, included: any, level: number): void {
+  private parseBelongsTo(data: any, included: Array<any>, remainingModels: Array<any>): void {
     const belongsTo: any = Reflect.getMetadata('BelongsTo', this);
 
     if (belongsTo) {
@@ -121,21 +169,22 @@ export class JsonApiModel {
           if (dataRelationship) {
             const typeName: string = dataRelationship.type;
             // tslint:disable-next-line:max-line-length
-            const modelType: ModelType<this> = Reflect.getMetadata('JsonApiDatastoreConfig', this._datastore.constructor).models[typeName];
+            const modelType: ModelType<this> = Reflect.getMetadata('JsonApiDatastoreConfig', this.internalDatastore.constructor).models[typeName];
+
             if (modelType) {
               const relationshipModel = this.getBelongsToRelationship(
                 modelType,
                 dataRelationship,
                 included,
                 typeName,
-                level
+                remainingModels
               );
 
               if (relationshipModel) {
                 this[metadata.propertyName] = relationshipModel;
               }
             } else {
-              throw { message: 'parseBelongsTo - Model type for relationship ' + typeName + ' not found.' };
+              throw { message: `parseBelongsTo - Model type for relationship ${typeName} not found.` };
             }
           }
         }
@@ -148,58 +197,65 @@ export class JsonApiModel {
     data: any,
     included: any,
     typeName: string,
-    level: number
+    remainingModels: Array<any>
   ): Array<T> {
     const relationshipList: Array<T> = [];
 
     data.forEach((item: any) => {
-      const relationshipData: any = find(included, { id: item.id, type: typeName });
+      const relationshipData: any = find(remainingModels, { id: item.id, type: typeName } as any);
 
       if (relationshipData) {
         const newObject: T = this.createOrPeek(modelType, relationshipData);
 
-        if (level <= 2) {
-          newObject.syncRelationships(relationshipData, included, level + 1);
-        }
+        const indexOfNewlyFoundModel = remainingModels.indexOf(relationshipData);
+        const modelsForProcessing = remainingModels.concat([]);
+        modelsForProcessing.splice(indexOfNewlyFoundModel, 1);
+
+        newObject.syncRelationships(relationshipData, included, modelsForProcessing);
+
         relationshipList.push(newObject);
       }
     });
+
     return relationshipList;
   }
-
 
   private getBelongsToRelationship<T extends JsonApiModel>(
     modelType: ModelType<T>,
     data: any,
-    included: any,
+    included: Array<any>,
     typeName: string,
-    level: number
+    remainingModels: Array<any>
   ): T | null {
     const id: string = data.id;
-    const relationshipData: any = find(included, { id, type: typeName });
+
+    const relationshipData: any = find(remainingModels, { id, type: typeName } as any);
 
     if (relationshipData) {
       const newObject: T = this.createOrPeek(modelType, relationshipData);
 
-      if (level <= 2) {
-        newObject.syncRelationships(relationshipData, included, level + 1);
-      }
+      const indexOfNewlyFoundModel = remainingModels.indexOf(relationshipData);
+      const modelsForProcessing = remainingModels.concat([]);
+      modelsForProcessing.splice(indexOfNewlyFoundModel, 1);
+
+      newObject.syncRelationships(relationshipData, included, modelsForProcessing);
 
       return newObject;
     }
-    return this._datastore.peekRecord(modelType, id);
+
+    return this.internalDatastore.peekRecord(modelType, id);
   }
 
   private createOrPeek<T extends JsonApiModel>(modelType: ModelType<T>, data: any): T {
-    const peek = this._datastore.peekRecord(modelType, data.id);
+    const peek = this.internalDatastore.peekRecord(modelType, data.id);
 
     if (peek) {
-      _.extend(peek, data.attributes);
+      _.extend(peek, this.internalDatastore.transformSerializedNamesToPropertyNames(modelType, data.attributes));
       return peek;
     }
 
-    const newObject: T = new modelType(this._datastore, data);
-    this._datastore.addToStore(newObject);
+    const newObject: T = this.internalDatastore.deserializeModel(modelType, data);
+    this.internalDatastore.addToStore(newObject);
 
     return newObject;
   }