+++ test/models/author.model.ts
@@ -4,6 +4,9 @@ import { JsonApiModel } from '../../src/models/json-api.model';
 import { Attribute } from '../../src/decorators/attribute.decorator';
 import { HasMany } from '../../src/decorators/has-many.decorator';
 import { PageMetaData } from './page-meta-data';
+import { JsonModelConverter} from '../../src/converters/json-model/json-model.converter';
+import { NestedAttribute } from '../../src/decorators/nested-attribute.decorator';
+import { School } from './school.model';
 
 @JsonApiModelConfig({
   type: 'authors',
@@ -24,6 +27,12 @@ export class Author extends JsonApiModel {
   @Attribute()
   updated_at: Date;
 
+  @Attribute()
+  firstNames: string[];
+
   @HasMany()
   books: Book[];
+
+  @NestedAttribute({ converter: new JsonModelConverter<any>(School) })
+  school: School;
 }