+++ CHANGELOG.md
@@ -1,3 +1,48 @@
+# [7.0.0]
+
+  ### BREAKING CHANGES
+
+  * Upgraded the library to Angular 7
+
+# [6.1.3]
+
+  * Fixed rollbackAttributes to public
+  * Fixed tests
+
+# [6.1.2]
+
+  * Added support for custom request options
+
+# [6.1.2-beta]
+
+  * Added support for custom request options
+
+# [6.0.1]
+
+  * Angular 6 support
+  * Smaller bug fixes
+
+### BREAKING CHANGES
+
+  * Required RxJS v6
+
+# [6.0.2-beta]
+
+  * Fix types issues
+
+# [6.0.0-beta]
+
+### BREAKING CHANGES
+
+  * Upgraded rxjs to version 6
+
+# [5.1.1-beta], [5.1.1] (2018-06-06)
+
+### Bug fixes
+
+  * Fix serializedName in included relationships ([#174](https://github.com/ghidoz/angular2-jsonapi/issues/174))
+  * Fix parsing belongsTo and hasMany relationships
+
 # [5.0.0] (2018-11-04)
 
 ### BREAKING CHANGES